Move init and shutdown stuff out of editor.cpp
