+++ qtfred/src/main.cpp
@@ -17,6 +17,7 @@
 #include "globalincs/globals.h" // MAX_SHIPS, NAME_LENGTH
 
 #include "mission/editor.h"
+#include "mission/management.h"
 
 #include "ui/FredView.h"
 #include "FredApplication.h"
