+++ qtfred/src/mission/editor.cpp
@@ -17,28 +17,13 @@
 #include <gamesnd/eventmusic.h>
 #include <starfield/nebula.h>
 
-#include "globalincs/pstypes.h" // vm_init
-#include "osapi/osregistry.h" // os_registry_init
-#include "io/timer.h" // timer_init
-#include "cfile/cfile.h" // cfile_chdir, cfile_init
-#include "localization/localize.h" // lcl_init
-#include "graphics/2d.h" // gr_init
-#include "graphics/font.h" // gr_font_init
-#include "io/key.h" // key_init
-#include "io/mouse.h" // mouse_init
 #include "iff_defs/iff_defs.h" // iff_init
 #include "object/object.h" // obj_init
 #include "species_defs/species_defs.h" // species_init
-#include "mission/missionbriefcommon.h" // mission_brief_init
-#include "ai/ai.h" // ai_init
-#include "ai/ai_profiles.h" // ai_profiles_init
 #include "ship/ship.h" // armor_init, ship_init
 #include "weapon/weapon.h" // weapon_init
-#include "stats/medals.h" // parse_medal_tbl
 #include "nebula/neb.h" // neb2_init
 #include "starfield/starfield.h" // stars_init, stars_pre_level_init, stars_post_level_init
-#include "osapi/osapi.h" // os_get_window, os_set_window.
-#include "graphics/font.h"
 #include "hud/hudsquadmsg.h"
 #include "globalincs/linklist.h"
 
@@ -46,99 +31,21 @@
 #include "ai/aigoals.h"
 
 #include "object.h"
-#include <FredApplication.h>
+#include "management.h"
+#include "FredApplication.h"
 
-extern int Xstr_inited;
+char Fred_callsigns[MAX_SHIPS][NAME_LENGTH + 1];
+char Fred_alt_names[MAX_SHIPS][NAME_LENGTH + 1];
+
+extern void allocate_mission_text(size_t size);
 
 extern int Nmodel_num;
 extern int Nmodel_instance_num;
 extern matrix Nmodel_orient;
 extern int Nmodel_bitmap;
 
-extern void allocate_mission_text(size_t size);
-
-extern void parse_init(bool basic = false);
-
-char Fred_callsigns[MAX_SHIPS][NAME_LENGTH + 1];
-char Fred_alt_names[MAX_SHIPS][NAME_LENGTH + 1];
-
 namespace fso {
 namespace fred {
-
-void initialize(const std::string& cfilepath, int argc, char* argv[], Editor* editor, InitializerCallback listener) {
-
-#ifndef NDEBUG
-	outwnd_init();
-
-	load_filter_info();
-#endif
-
-	std::setlocale(LC_ALL, "C");
-
-	std::vector<std::pair<std::function<void(void)>, SubSystem>> initializers =
-		{{ std::bind(os_init, Osreg_company_name, Osreg_app_name, nullptr, nullptr), SubSystem::OS },
-		 { timer_init,                                                               SubSystem::Timer },
-		 { [&]() {
-			 // this should enable mods - Kazan
-			 parse_cmdline(argc, argv);
-		 },                                                                          SubSystem::CommandLine },
-		 { [&cfilepath](void) {
-			 // cfile needs a file path...
-			 auto file_path = cfilepath + DIR_SEPARATOR_STR + "qtfred.exe";
-
-			 cfile_init(file_path.c_str());
-		 },                                                                          SubSystem::CFile },
-		 { []() {
-			 lcl_init(FS2_OPEN_DEFAULT_LANGUAGE);
-
-			 // Goober5000 - force init XSTRs (so they work, but only work in English, based on above comment)
-			 Xstr_inited = 1;
-		 },                                                                          SubSystem::Locale },
-		 { [=]() {
-			 std::unique_ptr<QtGraphicsOperations> graphicsOps(new QtGraphicsOperations(editor));
-			 gr_init(std::move(graphicsOps));
-		 },                                                                          SubSystem::Graphics },
-		 { font::init,                                                               SubSystem::Fonts },
-		 { key_init,                                                                 SubSystem::Keyboard },
-		 { mouse_init,                                                               SubSystem::Mouse },
-		 { particle::ParticleManager::init,                                          SubSystem::Particles },
-		 { iff_init,                                                                 SubSystem::Iff },
-		 { obj_init,                                                                 SubSystem::Objects },
-		 { species_init,                                                             SubSystem::Species },
-		 { mission_brief_common_init,                                                SubSystem::MissionBrief },
-		 { ai_init,                                                                  SubSystem::AI },
-		 { ai_profiles_init,                                                         SubSystem::AIProfiles },
-		 { armor_init,                                                               SubSystem::Armor },
-		 { weapon_init,                                                              SubSystem::Weapon },
-		 { parse_medal_tbl,                                                          SubSystem::Medals },
-		 { ship_init,                                                                SubSystem::Ships },
-		 { []() { parse_init(); },                                                   SubSystem::Parse },
-		 { neb2_init,                                                                SubSystem::Nebulas },
-		 { []() {
-			 stars_init();
-			 stars_pre_level_init(true);
-			 stars_post_level_init();
-		 },                                                                          SubSystem::Stars }};
-
-	for (const auto& initializer : initializers) {
-		const auto& init_function = initializer.first;
-		const auto& which = initializer.second;
-		listener(which);
-		init_function();
-	}
-}
-
-void shutdown() {
-	audiostream_close();
-	snd_close();
-	gr_close();
-
-	os_cleanup();
-
-#ifndef NDEBUG
-	outwnd_close();
-#endif
-}
 
 Editor::Editor() : currentObject{ -1 } {
 	// We need to do this the hard way since MSVC 2013 is too stupid to support array initializers...
